+++ editor/src/com/kotcrab/vis/editor/module/scene/system/inflater/EditorSoundInflater.java
@@ -18,6 +18,11 @@ package com.kotcrab.vis.editor.module.scene.system.inflater;
 
 import com.artemis.Aspect;
 import com.artemis.ComponentMapper;
+import com.kotcrab.vis.editor.Log;
+import com.kotcrab.vis.editor.module.project.FileAccessModule;
+import com.kotcrab.vis.editor.module.scene.AssetsLoadingMonitorModule;
+import com.kotcrab.vis.runtime.assets.SoundAsset;
+import com.kotcrab.vis.runtime.assets.VisAssetDescriptor;
 import com.kotcrab.vis.runtime.component.AssetReference;
 import com.kotcrab.vis.runtime.component.VisSound;
 import com.kotcrab.vis.runtime.component.proto.ProtoVisSound;
@@ -25,6 +30,10 @@ import com.kotcrab.vis.runtime.system.inflater.InflaterSystem;
 
 /** @author Kotcrab */
 public class EditorSoundInflater extends InflaterSystem {
+	private FileAccessModule fileAccessModule;
+	private AssetsLoadingMonitorModule loadingMonitor;
+
+	private ComponentMapper<AssetReference> assetCm;
 	private ComponentMapper<ProtoVisSound> protoCm;
 	private ComponentMapper<VisSound> soundCm;
 
@@ -34,6 +43,20 @@ public class EditorSoundInflater extends InflaterSystem {
 
 	@Override
 	public void inserted (int entityId) {
+		VisAssetDescriptor assetDescriptor = assetCm.get(entityId).asset;
+
+		if (assetDescriptor instanceof SoundAsset) {
+			String path = ((SoundAsset) assetDescriptor).getPath();
+			boolean exists = fileAccessModule.getAssetsFolder().child(path).exists();
+			if (exists == false) {
+				String errorMsg = "Sound file does not exist: " + path;
+				Log.fatal(errorMsg);
+				loadingMonitor.addFailedResource(assetDescriptor, new IllegalStateException(errorMsg));
+			}
+		} else {
+			throw new IllegalStateException("Unsupported asset descriptor for sound inflater: " + assetDescriptor);
+		}
+
 		soundCm.create(entityId);
 		protoCm.remove(entityId);
 	}