+++ editor/src/com/kotcrab/vis/editor/module/scene/system/inflater/EditorSpriteInflater.java
@@ -18,7 +18,13 @@ package com.kotcrab.vis.editor.module.scene.system.inflater;
 
 import com.artemis.Aspect;
 import com.artemis.ComponentMapper;
+import com.badlogic.gdx.graphics.g2d.TextureRegion;
+import com.badlogic.gdx.utils.GdxRuntimeException;
+import com.kotcrab.vis.editor.Icons;
+import com.kotcrab.vis.editor.Log;
 import com.kotcrab.vis.editor.module.project.TextureCacheModule;
+import com.kotcrab.vis.editor.module.scene.AssetsLoadingMonitorModule;
+import com.kotcrab.vis.runtime.assets.VisAssetDescriptor;
 import com.kotcrab.vis.runtime.component.AssetReference;
 import com.kotcrab.vis.runtime.component.VisSprite;
 import com.kotcrab.vis.runtime.component.proto.ProtoVisSprite;
@@ -26,10 +32,12 @@ import com.kotcrab.vis.runtime.system.inflater.InflaterSystem;
 
 /** @author Kotcrab */
 public class EditorSpriteInflater extends InflaterSystem {
+	private TextureCacheModule textureCache;
+	private AssetsLoadingMonitorModule loadingMonitor;
+
 	private ComponentMapper<VisSprite> spriteCm;
 	private ComponentMapper<ProtoVisSprite> protoCm;
 	private ComponentMapper<AssetReference> assetCm;
-	private TextureCacheModule textureCache;
 
 	public EditorSpriteInflater () {
 		super(Aspect.all(ProtoVisSprite.class, AssetReference.class));
@@ -37,8 +45,19 @@ public class EditorSpriteInflater extends InflaterSystem {
 
 	@Override
 	public void inserted (int entityId) {
+		VisAssetDescriptor asset = assetCm.get(entityId).asset;
+		TextureRegion region;
+
+		try {
+			region = textureCache.getRegion(asset);
+		} catch (GdxRuntimeException e) {
+			Log.exception(e);
+			region = Icons.QUESTION_BIG.textureRegion();
+			loadingMonitor.addFailedResource(asset, e);
+		}
+
 		VisSprite sprite = spriteCm.create(entityId);
-		sprite.setRegion(textureCache.getRegion(assetCm.get(entityId).asset));
+		sprite.setRegion(region);
 		protoCm.get(entityId).fill(sprite);
 		protoCm.remove(entityId);
 	}